fix: 구조변경(추상 클래스, 구현 클래스 분리)
