+++ API_SERVICE/common_service/routes/v1/select.py
@@ -5,7 +5,7 @@ from fastapi import Depends, APIRouter
 from pydantic import BaseModel
 
 from common_service.database.conn import db
-from libs.database.conn import Connector
+from libs.database.connector import Connector
 
 
 class JoinInfo(BaseModel):
