+++ API_SERVICE/login_service/database/conn.py
@@ -1,8 +1,8 @@
 from sqlalchemy.orm import declarative_base
 
 from login_service.common.config import settings
-from libs.database.conn import TiberoConnector
-from libs.database.conn import SQLAlchemy
+from libs.database.tibero import TiberoConnector
+from libs.database.orm import SQLAlchemyConnector
 
 Base = declarative_base()
-db = SQLAlchemy(Base) if settings.DB_INFO.type != "tibero" else TiberoConnector()
+db = SQLAlchemyConnector(Base) if settings.DB_INFO.type != "tibero" else TiberoConnector()
