+++ common_libs/libs/database/connector.py
@@ -1,0 +1,36 @@
+import abc
+from typing import TypeVar, Tuple, List, Dict
+
+from fastapi import FastAPI
+
+T = TypeVar('T', bound="Connector")
+
+
+class Connector(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def init_app(self, app: FastAPI, **kwargs):
+        ...
+
+    @abc.abstractmethod
+    def get_db(self) -> T:
+        ...
+
+    @abc.abstractmethod
+    def query(self, **kwargs) -> T:
+        ...
+
+    @abc.abstractmethod
+    def all(self) -> Tuple[List[dict], int]:
+        ...
+
+    @abc.abstractmethod
+    def first(self) -> dict:
+        ...
+
+    @abc.abstractmethod
+    def execute(self, tablename, **kwargs):
+        ...
+
+    @abc.abstractmethod
+    def get_column_info(self, table_nm) -> List[Dict[str, str]]:
+        ...
